move: domain - application 에서 application으로의 파일 이동
