+++ src/main/java/org/pokeherb/productservice/application/command/ProductCommandServiceImpl.java
@@ -1,11 +1,11 @@
-package org.pokeherb.productservice.product.domain.application.command;
+package org.pokeherb.productservice.application.command;
 
 import lombok.RequiredArgsConstructor;
 import org.pokeherb.productservice.global.infrastructure.client.VendorServiceClient;
 import org.pokeherb.productservice.global.infrastructure.exception.CustomException;
 import org.pokeherb.productservice.product.domain.ProductRepository;
 import org.pokeherb.productservice.product.domain.VendorHubIdDto;
-import org.pokeherb.productservice.product.domain.application.dto.ProductDto;
+import org.pokeherb.productservice.application.dto.ProductDto;
 import org.pokeherb.productservice.product.domain.entity.Product;
 import org.pokeherb.productservice.product.domain.exception.ProductErrorCode;
 import org.pokeherb.productservice.product.presentation.dto.ProductCreateRequestDto;
