+++ src/main/java/org/pokeherb/productservice/product/presentation/ProductController.java
@@ -3,8 +3,8 @@ package org.pokeherb.productservice.product.presentation;
 import lombok.RequiredArgsConstructor;
 import org.pokeherb.productservice.global.infrastructure.CustomResponse;
 import org.pokeherb.productservice.global.infrastructure.success.GeneralSuccessCode;
-import org.pokeherb.productservice.product.domain.application.command.ProductCommandService;
-import org.pokeherb.productservice.product.domain.application.dto.ProductDto;
+import org.pokeherb.productservice.application.command.ProductCommandService;
+import org.pokeherb.productservice.application.dto.ProductDto;
 import org.pokeherb.productservice.product.presentation.dto.ProductCreateRequestDto;
 import org.pokeherb.productservice.product.presentation.dto.ProductUpdateRequestDto;
 import org.springframework.web.bind.annotation.*;
